refactored class, updated template to reflect correct username
