+++ servers.d/server-template.ini
@@ -4,6 +4,6 @@ excludes = /proc,/bin,/boot,/dev,/lib,/lib64,/lost+found,/media,/mnt,/run,/sbin,
 
 [connection]
 host = server01.domain.com 
-user = root
+username = root
 port = 22
 ssh_args = -o UserKnownHostsFile=/dev/null,-o StrictHostKeyChecking=no
